backend/requirements.txt
flask>=3.0.0
flask-cors>=6.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyVCF3>=1.0.3
chardet>=5.0.0
charset-normalizer>=3.0.0
reportlab>=4.0.0
